#348 Extracted RichText sanitiser logic into RichTextSanitiser.
