+++ platform-pojo-bl/src/test/java/ua/com/fielden/platform/types/RichTextHtmlSanitisationTest.java
@@ -102,18 +102,18 @@ public class RichTextHtmlSanitisationTest {
     }
 
     private void assertSanitizationFailure(final String input) {
-        assertFalse(RichText.sanitiseMarkdown(input).isSuccessful());
+        assertFalse(RichTextSanitiser.sanitiseMarkdown(input).isSuccessful());
     }
 
     private void assertSanitizationSuccess(final String input) {
-        final Result result = RichText.sanitiseMarkdown(input);
+        final Result result = RichTextSanitiser.sanitiseMarkdown(input);
         if (!result.isSuccessful()) {
             fail(result.getMessage());
         }
     }
 
     private void assertAfterSanitization(final String expected, final String input) {
-        final Result result = RichText.sanitiseMarkdown(input);
+        final Result result = RichTextSanitiser.sanitiseMarkdown(input);
         if (!result.isSuccessful()) {
             fail(result.getMessage());
         } else {
