+++ platform-pojo-bl/src/test/java/ua/com/fielden/platform/types/RichTextMarkdownSanitisationTest.java
@@ -142,18 +142,18 @@ public class RichTextMarkdownSanitisationTest {
     }
 
     private void assertSanitizationFailure(final String input) {
-        assertFalse(RichText.sanitiseMarkdown(input).isSuccessful());
+        assertFalse(RichTextSanitiser.sanitiseMarkdown(input).isSuccessful());
     }
 
     private void assertSanitizationSuccess(final String input) {
-        final Result result = RichText.sanitiseMarkdown(input);
+        final Result result = RichTextSanitiser.sanitiseMarkdown(input);
         if (!result.isSuccessful()) {
             fail(result.getMessage());
         }
     }
 
     private void assertAfterSanitization(final String expected, final String input) {
-        final Result result = RichText.sanitiseMarkdown(input);
+        final Result result = RichTextSanitiser.sanitiseMarkdown(input);
         if (!result.isSuccessful()) {
             fail(result.getMessage());
         } else {
